requirements.txt
flask==2.0.1
numpy==1.22.4
opencv-python==4.6.0.66
pillow==9.2.0
gunicorn==20.1.0